+++ src/SaveBackup/ModEntry.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.IO.Compression;
@@ -114,91 +115,74 @@ public class ModEntry : MelonMod
     /// <summary>Back up the current saves and prune older backups as needed.</summary>
     private void UpdateBackups()
     {
-        if (this.Config is { IncludeSurvival: false, IncludeWintermute: false })
-        {
-            this.Log.Warning("Can't create backups: you disabled both survival and Wintermute backups, so there's nothing to do.");
-            return;
-        }
-
-        string backupLabel = $"MelonLoader {this.MelonLoaderVersion}, The Long Dark {this.GameVersion}";
-        if (this.Config.IncludeSurvival != this.Config.IncludeWintermute)
-        {
-            backupLabel += this.Config.IncludeSurvival
-                ? ", survival only"
-                : ", Wintermute only";
-        }
-
+        string backupLabel = this.GetBackupLabel();
         DateTimeOffset now = DateTimeOffset.Now;
 
-        this.UpdateBackupsOfType(
-            rootBackupsPath: this.BackupFolder,
-            type: DailyFolderName,
-            name: $"{now:yyyy-MM-dd} ({backupLabel})",
-            backupsToKeep: this.Config.DailyBackupCount
-        );
+        if (this.Config.DailyBackupCount > 0)
+            this.UpdateBackupsOfType(DailyFolderName, $"{now:yyyy-MM-dd} ({backupLabel})");
+        if (this.Config.HourlyBackupCount > 0)
+            this.UpdateBackupsOfType(HourlyFolderName, $"{now:yyyy-MM-dd HH} ({backupLabel})");
 
-        this.UpdateBackupsOfType(
-            rootBackupsPath: this.BackupFolder,
-            type: HourlyFolderName,
-            name: $"{now:yyyy-MM-dd HH} ({backupLabel})",
-            backupsToKeep: this.Config.HourlyBackupCount
-        );
+        this.PruneBackups(DailyFolderName, this.Config.DailyBackupCount);
+        this.PruneBackups(HourlyFolderName, this.Config.HourlyBackupCount);
+    }
+
+    /// <summary>Get the name for a backup taken now (excluding the data).</summary>
+    private string GetBackupLabel()
+    {
+        string backupLabel = $"MelonLoader {this.MelonLoaderVersion}, The Long Dark {this.GameVersion}";
+
+        switch (this.Config)
+        {
+            case { IncludeSurvival: true, IncludeWintermute: false }:
+                backupLabel += ", survival only";
+                break;
+
+            case { IncludeSurvival: false, IncludeWintermute: true }:
+                backupLabel += ", Wintermute only";
+                break;
+
+            case { IncludeSurvival: false, IncludeWintermute: false }:
+                backupLabel += ", mod data only";
+                break;
+        }
+
+        return backupLabel;
     }
 
     /// <summary>Back up the current saves of a given type and prune its older backups as needed.</summary>
-    /// <param name="rootBackupsPath">The root folder in which to store backup subfolders.</param>
     /// <param name="type">The backup type (e.g. <see cref="DailyFolderName"/>).</param>
     /// <param name="name">The unique name for the backup, excluding the path and extension.</param>
-    /// <param name="backupsToKeep">The number of backups to keep of this type.</param>
-    private void UpdateBackupsOfType(string rootBackupsPath, string type, string name, int backupsToKeep)
+    private void UpdateBackupsOfType(string type, string name)
     {
         try
         {
-            string folderPath = Path.Combine(rootBackupsPath, type);
+            // get paths
+            string backupRoot = Path.Combine(this.BackupFolder, type);
+            FileInfo targetFile = new(Path.Combine(backupRoot, $"{name}.zip"));
+            DirectoryInfo fallbackDir = new(Path.Combine(backupRoot, name));
 
-            // add backup
-            if (backupsToKeep > 0)
+            // skip: nothing to do
+            if (targetFile.Exists || fallbackDir.Exists || !this.TryGetFilesToBackUp(out Dictionary<string, FileInfo> backupFiles))
+                return;
+
+            // copy files into backup folder
+            foreach ((string relativePath, FileInfo file) in backupFiles)
             {
-                Directory.CreateDirectory(folderPath);
+                string copyToPath = Path.Combine(fallbackDir.FullName, relativePath);
 
-                // get target path
-                FileInfo targetFile = new(Path.Combine(folderPath, $"{name}.zip"));
-                DirectoryInfo fallbackDir = new(Path.Combine(folderPath, name));
-                if (targetFile.Exists || fallbackDir.Exists)
-                    return;
-
-                // copy saves to fallback directory
-                DirectoryInfo savesDir = new(this.SavesPath);
-                bool copiedAny = false;
-
-                foreach (FileSystemInfo entry in savesDir.GetFileSystemInfos())
-                {
-                    bool isSurvival = entry is DirectoryInfo && entry.Name.Equals("Survival", StringComparison.OrdinalIgnoreCase);
-                    bool include = isSurvival
-                        ? this.Config.IncludeSurvival
-                        : this.Config.IncludeWintermute;
-
-                    if (!include)
-                        continue;
-
-                    copiedAny |= this.RecursiveCopy(entry, fallbackDir);
-                }
-
-                // compress backup if possible
-                if (copiedAny)
-                {
-                    if (!this.TryCompressDir(fallbackDir.FullName, targetFile, out Exception? compressError))
-                        this.Log.Msg($"Added {type} backup at {fallbackDir.FullName}. Couldn't compress backup:\n{compressError}");
-                    else
-                    {
-                        this.Log.Msg($"Added {type} backup at {targetFile.FullName}.");
-                        fallbackDir.Delete(recursive: true);
-                    }
-                }
+                Directory.CreateDirectory(Path.GetDirectoryName(copyToPath)!);
+                File.Copy(file.FullName, copyToPath);
             }
 
-            // prune backups
-            this.PruneBackups(type, new DirectoryInfo(folderPath), backupsToKeep);
+            // zip & delete folder
+            if (!this.TryCompressDir(fallbackDir.FullName, targetFile, out Exception? compressError))
+                this.Log.Msg($"Added {type} backup at {fallbackDir.FullName}. Couldn't compress backup:\n{compressError}");
+            else
+            {
+                this.Log.Msg($"Added {type} backup at {targetFile.FullName}.");
+                fallbackDir.Delete(recursive: true);
+            }
         }
         catch (Exception ex)
         {
@@ -206,14 +190,54 @@ public class ModEntry : MelonMod
         }
     }
 
+    /// <summary>Get the files that should be backed up, indexed by their relative path within the backup.</summary>
+    /// <param name="files">The files to back up.</param>
+    /// <returns>Returns whether any files were found.</returns>
+    private bool TryGetFilesToBackUp(out Dictionary<string, FileInfo> files)
+    {
+        files = [];
+
+        // save files
+        DirectoryInfo savesDir = new(this.SavesPath);
+        if (savesDir.Exists)
+        {
+            foreach (FileSystemInfo entry in savesDir.GetFileSystemInfos())
+            {
+                bool isSurvival = entry is DirectoryInfo && entry.Name.Equals("Survival", StringComparison.OrdinalIgnoreCase);
+                bool include = isSurvival ? this.Config.IncludeSurvival : this.Config.IncludeWintermute;
+                if (!include)
+                    continue;
+
+                foreach (FileInfo file in this.GetFilesRecursively(entry))
+                {
+                    string relativePath = Path.GetRelativePath(savesDir.FullName, file.FullName);
+                    files[Path.Combine("Saves", relativePath)] = file;
+                }
+            }
+        }
+
+        // mod data
+        DirectoryInfo modDataDir = new DirectoryInfo(Path.Combine(MelonEnvironment.ModsDirectory, "ModData"));
+        foreach (FileInfo file in this.GetFilesRecursively(modDataDir))
+        {
+            string relativePath = Path.GetRelativePath(MelonEnvironment.ModsDirectory, file.FullName);
+            files[relativePath] = file;
+        }
+
+        return files.Count > 0;
+    }
+
     /// <summary>Remove old backups if we've exceeded the limit.</summary>
     /// <param name="type">The backup type (e.g. <see cref="DailyFolderName"/>).</param>
-    /// <param name="backupFolder">The folder containing save backups of the type being pruned.</param>
     /// <param name="backupsToKeep">The number of backups to keep.</param>
-    private void PruneBackups(string type, DirectoryInfo backupFolder, int backupsToKeep)
+    private void PruneBackups(string type, int backupsToKeep)
     {
         try
         {
+            DirectoryInfo backupFolder = new DirectoryInfo(
+                Path.Combine(this.BackupFolder, type)
+            );
+
             if (!backupFolder.Exists)
                 return;
 
@@ -265,35 +289,23 @@ public class ModEntry : MelonMod
         }
     }
 
-    /// <summary>Recursively copy a directory or file.</summary>
-    /// <param name="source">The file or folder to copy.</param>
-    /// <param name="targetFolder">The folder to copy it into.</param>
-    /// <returns>Returns whether any files were copied.</returns>
-    private bool RecursiveCopy(FileSystemInfo source, DirectoryInfo targetFolder)
+    /// <summary>Recursively get all files starting from a root (including the root itself if it's a file).</summary>
+    /// <param name="root">The file or directory from which to start.</param>
+    private FileInfo[] GetFilesRecursively(FileSystemInfo root)
     {
-        if (!source.Exists)
-            return false;
+        if (!root.Exists)
+            return [];
 
-        bool anyCopied = false;
-
-        switch (source)
+        switch (root)
         {
             case FileInfo sourceFile:
-                targetFolder.Create();
-                sourceFile.CopyTo(Path.Combine(targetFolder.FullName, sourceFile.Name));
-                anyCopied = true;
-                break;
+                return [sourceFile];
 
             case DirectoryInfo sourceDir:
-                DirectoryInfo targetSubfolder = new DirectoryInfo(Path.Combine(targetFolder.FullName, sourceDir.Name));
-                foreach (var entry in sourceDir.EnumerateFileSystemInfos())
-                    anyCopied = this.RecursiveCopy(entry, targetSubfolder) || anyCopied;
-                break;
+                return sourceDir.GetFiles("*.*", SearchOption.AllDirectories);
 
             default:
-                throw new NotSupportedException($"Unknown filesystem info type '{source.GetType().FullName}'.");
+                throw new NotSupportedException($"Unknown filesystem info type '{root.GetType().FullName}'.");
         }
-
-        return anyCopied;
     }
 }