+++ src/SaveBackup/docs/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Backups now include mod save data (from `Mods/ModData/*`).
+* Backups now include a `README.txt` file explaining how to restore them.
 
 ## 1.2.0
 Released 22 April 2026 for The Long Dark 2.50+.